+++ src/split.jl
@@ -30,15 +30,17 @@ findwhen(ta::TimeArray{Bool,1}) = ta.timestamp[find(ta.values)]
 ###### head, tail ###########
 
 function head{T,N,D}(ta::TimeArray{T,N,D}, n::Int=1)
+    ncol          = length(ta.colnames)
     new_timestamp = ta.timestamp[1:n]
-    new_values    = ta.values[1:n]
+    new_values    = ta.values[1:n, 1:ncol]
     TimeArray(new_timestamp, new_values, ta.colnames, ta.meta)
 end
 
 function tail{T,N,D}(ta::TimeArray{T,N,D}, n::Int=1)
+    ncol          = length(ta.colnames)
     tail_start = length(ta)-n+1
     new_timestamp = ta.timestamp[tail_start:end]
-    new_values    = ta.values[tail_start:end]
+    new_values    = ta.values[tail_start:end, 1:ncol]
     TimeArray(new_timestamp, new_values, ta.colnames, ta.meta)
 end